+++ transport/src/main/java/com/cooperative/transport/repositories/UtilisateurRepository.java
@@ -9,10 +9,10 @@ import org.springframework.stereotype.Repository;
 
 import com.cooperative.transport.entities.Utilisateurs;
 @Repository
-public interface UtilisateurRepository extends JpaRepository<Utilisateurs, Long> {
+public interface UtilisateurRepository extends JpaRepository<Utilisateurs, Integer> {
     @Query("SELECT DISTINCT e,s,r FROM Employes e LEFT JOIN e.salaires s JOIN e.role r where s.date_modification = (SELECT MAX(s2.date_modification) FROM Salaires s2 WHERE s2.employe.id = e.id ) or s is null")
     List<Object[]> findEmploye();
     @Query("SELECT DISTINCT e,s,r FROM Employes e LEFT JOIN e.salaires s JOIN e.role r where (s.date_modification = (SELECT MAX(s2.date_modification) FROM Salaires s2 WHERE s2.employe.id = e.id ) or s is null) and e.id = :id")
-    List<Object[]> findEmployeById(Long id);
+    List<Object[]> findEmployeById(Integer id);
     List<Utilisateurs> findAllChauffeurDispo(LocalDateTime dateCible);
 }
